Supports TS setup without wildcards in packages
With this commit, the plugin supports a TS setup where

* Workspaces are configured by name (['packages/a', 'packages/b]'), not
  with a wildcard (['packages/*'])
* TS subprojects where the rootDir is '.'

To support the first feature, when the workspace entry does not contain
a wild card, only a single directory is processed instead of a list
of directories.

To support the second feature, the plugin now also remembers the package
directory for each file. Then, when the rootDir entry from the tsconfig
is '.', it can use this package directory to correctly get the output
directory.